+++ src/views/InitializationView/InitializationView.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import InitializationForm from './InitializationForm/InitializationForm';
 
 function InitializationView(): JSX.Element {
-  return <InitializationForm />;
+  return (
+    <div style={{ paddingLeft: '10px' }}>
+      <InitializationForm />
+    </div>
+  );
 }
 
 export default InitializationView;