+++ CHANGELOG.md
@@ -13,6 +13,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Rent increase
     - Generate notification document on rent increase
 
+### Changed
+
+- Initialization form
+    - Renamed last tab to `Others` instead of `Water costs`
+
 ## v0.1.1 - 2024-08-11
 
 ### Changed